requirements.txt
tensorflow==2.4.1
pandas==1.1.3
seaborn==0.11.0
streamlit==0.81.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.2.0
scikit_learn==0.24.2
